+++ lib/enums/pos/align_mode_enum.dart
@@ -1,0 +1,9 @@
+/// Enum para os valores de alinhamento
+enum AlignModeEnum {
+  left("LEFT"),
+  center("CENTER"),
+  right("RIGHT");
+
+  final String value;
+  const AlignModeEnum(this.value);
+}