+++ lib/enums/pos/font_format_enum.dart
@@ -1,0 +1,9 @@
+/// Enum para os formatos de fonte
+enum FontFormatEnum {
+  small("SMALL"),
+  medium("MEDIUM"),
+  large("LARGE");
+
+  final String value;
+  const FontFormatEnum(this.value);
+}